update a few thingy

+++ HELLWASHER/HELLWAHSER_Controller/Controllers/FeedbackController.cs
@@ -1,8 +1,10 @@
 ﻿using BusinessObject.IService;
 using BusinessObject.ViewModels.Feedback;
 using DataAccess;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Threading.Tasks;
 
 namespace HELLWASHER_Controller.Controllers
 {
@@ -12,59 +14,93 @@ namespace HELLWASHER_Controller.Controllers
     {
         private readonly IFeedbackService _feedbackService;
         private readonly WashShopContext _context;
-        public FeedbackController(IFeedbackService feedbackService, WashShopContext context)
+        private readonly IUserService _userService;
+
+        public FeedbackController(IFeedbackService feedbackService, WashShopContext context, IUserService userService)
         {
             _feedbackService = feedbackService;
             _context = context;
+            _userService = userService;
         }
+
         [HttpGet]
+        [Authorize(Roles = "Admin, Staff, Customer")]
         public async Task<IActionResult> GetAllFeedbacks()
         {
+            var user = await _userService.GetUserByTokenAsync(User);
+            if (user == null)
+            {
+                return Unauthorized();
+            }
+
             var result = await _feedbackService.GetAllFeedback();
             if (!result.Success) return BadRequest(result);
 
             return Ok(result);
         }
+
         [HttpGet("{feedbackId}")]
+        [Authorize(Roles = "Admin, Staff, Customer")]
         public async Task<IActionResult> GetFeedbackById(int feedbackId)
         {
+            var user = await _userService.GetUserByTokenAsync(User);
+            if (user == null)
+            {
+                return Unauthorized();
+            }
+
             var result = await _feedbackService.GetFeedbackById(feedbackId);
             if (!result.Success) return BadRequest(result);
 
             return Ok(result);
         }
-        [HttpPost("Add/{userId}")]
-        public async Task<IActionResult> AddFeedback([FromForm] FeedbackRequest feedbackDTO, int userId)
+
+        [HttpPost("Add")]
+        [Authorize(Roles = "Customer")]
+        public async Task<IActionResult> AddFeedback([FromForm] FeedbackRequest feedbackDTO)
         {
-            var result = await _feedbackService.AddFeedback(feedbackDTO, userId);
+            var user = await _userService.GetUserByTokenAsync(User);
+            if (user == null)
+            {
+                return Unauthorized();
+            }
+
+            var result = await _feedbackService.AddFeedback(feedbackDTO, user.UserId);
             if (!result.Success) return BadRequest(result);
 
             return Ok(result);
         }
+
         [HttpPut("update/{feedbackId}")]
-        public async Task<IActionResult> UpdateFeedback([FromBody] FeedbackRequest feedbackRequest, int feedbackId)
+        [Authorize(Roles = "Customer")]
+        public async Task<IActionResult> UpdateFeedback([FromForm] FeedbackRequest feedbackRequest, int feedbackId)
         {
+            var user = await _userService.GetUserByTokenAsync(User);
+            if (user == null)
+            {
+                return Unauthorized();
+            }
+
             var result = await _feedbackService.UpdateFeedback(feedbackRequest, feedbackId);
             if (!result.Success) return BadRequest(result);
 
             return Ok(result);
         }
+
         [HttpDelete("delete/{feedbackId}")]
+        [Authorize(Roles = "Customer")]
         public async Task<IActionResult> DeleteFeedback(int feedbackId)
         {
+            var user = await _userService.GetUserByTokenAsync(User);
+            if (user == null)
+            {
+                return Unauthorized();
+            }
+
             var result = await _feedbackService.DeleteFeedback(feedbackId);
             if (!result.Success) return BadRequest(result);
 
             return Ok(result);
         }
-        //[HttpPatch("image")]
-        //public async Task<IActionResult> UpdateFeedbackImage(int feedbackId, IFormFile image)
-        //{
-        //    var result = await _feedbackService.UpdateFeedbackImage(feedbackId, image);
-        //    if (!result.Success) return BadRequest(result);
-
-        //    return Ok(result);
-        //}
-
     }
 }
